Rename: {email,oidc}.rs -> {email,oidc}_bridge.rs
Avoids clashing with handler namespaces introduced in future commits
