+++ src/lib.rs
@@ -20,10 +20,10 @@ pub mod error;
 pub mod config;
 pub use config::{Config, ConfigBuilder};
 pub mod crypto;
-pub mod email;
+pub mod email_bridge;
 pub mod middleware;
 pub mod handlers;
-pub mod oidc;
+pub mod oidc_bridge;
 pub mod store;
 pub mod store_cache;
 pub mod validation;
